+++ changelog.md
@@ -1,7 +1,7 @@
 InsaneLib's Changelog
 
 ## Upcoming
-* Refactored IdTagMatcher. No more matchesGeneric.
+* Refactored IdTagMatcher. No more matchesGeneric. Affects Blacklist too. Added more getAll methods.
 * Removed MCUtils.isBlockingWithShield as it's now public in LivingEntity
 * Removed RandomHelper, there's Mth